+++ ソースコード/9_アナログ入力/analog_input_kadai2/analog_input_kadai2.ino
@@ -18,7 +18,7 @@ void setup() {
 }
 
 void loop() {
-  float volt = analogRead(VOL)*((float)5/1024);
+  double volt = analogRead(VOL)*((double)5/1024);
   Serial.println(volt);
   
   if(volt >= 3.0){
